Loading fails if a yaml file contains the substring python.
A yaml file that contains the substring python is considered to be a python file due to the is_python_file function defined in dax_tools_utils.py. I've change the order in which the conditions are tested; first it is checked for a .yaml ending, then we check whether it has a .py ending or the substring python.
We shouldn't really be checking file types in this way, but at that is a separate and more extensive refactor.